+++ .pre-commit-config.yaml
@@ -6,6 +6,7 @@ repos:
     rev: v4.2.0
     hooks:
       - id: trailing-whitespace
+        exclude: docs/(source|build/html)/_static/
       - id: end-of-file-fixer
         exclude: docs/source/_static/
       - id: check-yaml
